Mudanças no seeder
Mudança do nome de 'WorflowSeeder' para 'WorkflowSeeder'. Além disso, o nome do primeiro formulário
agora consulta o arquivo de config ('app.initial_form') para setar o nome do formulário.
