+++ database/seeders/FormDefinitionsTableSeeder.php
@@ -9,8 +9,10 @@ class FormDefinitionsTableSeeder extends Seeder
 {
     public function run()
     {
+        $initial_form_name = config('app.initial_form');
+
         $formDefinitions = [
-            [1, 'tr_inicio_conferencia', 'equivalencia', null, '[ [ { "name": "unidade_ies", "type": "text", "label": "Unidade\/IES onde cursou a disciplina:", "required": true }, { "name": "hist_esc", "type": "file", "accept": ".pdf", "label": "Histórico Escolar:", "required": true, "width": 4 } ], [ { "name": "coddis1", "type": "text", "label": "Código 1", "required": true, "width": 2 }, { "name": "disciplina1", "type": "text", "label": "Disciplina 1", "required": true }, { "name": "file_dis1", "type": "file", "label": "Ementa da Disciplina 1", "accept": ".pdf", "required": true, "width": 4 } ], [ { "name": "semestre_dis1", "width": 3, "type": "select", "label": "Semestre em que cursou:", "options": [ "1°", "2°" ], "required": true }, { "name": "ano_dis1", "type": "text", "label": "Ano:", "required": true }, { "name": "freq_dis1", "width": 1, "type": "text", "label": "Frequência:", "required": true }, { "name": "nota_dis1", "width": 1, "type": "text", "label": "Nota:", "required": true }, { "name": "credit_dis1", "width": 1, "type": "text", "label": "Créditos", "required": true }, { "name": "cghr_dis1", "type": "text", "label": "Carga horária:", "required": true } ], [ { "name": "coddis2", "type": "text", "label": "Código 2", "required": false, "width": 2 }, { "name": "disciplina2", "type": "text", "label": "Disciplina 2", "required": false }, { "name": "file_dis2", "type": "file", "label": "Ementa da Disciplina 2", "accept": ".pdf", "required": false, "width": 4 } ], [ { "name": "semestre_dis2", "width": 3, "type": "select", "label": "Semestre em que cursou:", "options": [ "1°", "2°" ], "required": false }, { "name": "ano_dis2", "type": "text", "label": "Ano:", "required": false }, { "name": "freq_dis2", "width": 1, "type": "text", "label": "Frequência:", "required": false }, { "name": "nota_dis2", "width": 1, "type": "text", "label": "Nota:", "required": false }, { "name": "credit_dis2", "width": 1, "type": "text", "label": "Créditos", "required": false }, { "name": "cghr_dis2", "type": "text", "label": "Carga horária:", "required": false } ], [ { "name": "coddis3", "type": "text", "label": "Código 3", "required": false, "width": 2 }, { "name": "disciplina3", "type": "text", "label": "Disciplina 3", "required": false }, { "name": "file_dis3", "type": "file", "label": "Ementa da Disciplina 3", "accept": ".pdf", "required": false, "width": 4 } ], [ { "name": "semestre_dis3", "width": 3, "type": "select", "label": "Semestre em que cursou:", "options": [ "1°", "2°" ], "required": false }, { "name": "ano_dis3", "type": "text", "label": "Ano:", "required": false }, { "name": "freq_dis3", "width": 1, "type": "text", "label": "Frequência:", "required": false }, { "name": "nota_dis3", "width": 1, "type": "text", "label": "Nota:", "required": false }, { "name": "credit_dis3", "width": 1, "type": "text", "label": "Créditos", "required": false }, { "name": "cghr_dis3", "type": "text", "label": "Carga horária:", "required": false } ], [ { "name": "coddis4", "type": "disciplina-usp", "label": "Disciplina à requerir equivalência", "required": true, "width": 2 }, { "name": "disciplina4", "type": "text", "label": "Nome da disciplina desejada", "required": true } ], { "name": "place", "type": "hidden", "value": "place_name" }, { "name": "transition", "type": "hidden", "value": "transition_name" } ]', '2025-03-17 19:33:49', '2026-03-26 20:20:12'],
+            [1, $initial_form_name, 'equivalencia', null, '[ [ { "name": "unidade_ies", "type": "text", "label": "Unidade\/IES onde cursou a disciplina:", "required": true }, { "name": "hist_esc", "type": "file", "accept": ".pdf", "label": "Histórico Escolar:", "required": true, "width": 4 } ], [ { "name": "coddis1", "type": "text", "label": "Código 1", "required": true, "width": 2 }, { "name": "disciplina1", "type": "text", "label": "Disciplina 1", "required": true }, { "name": "file_dis1", "type": "file", "label": "Ementa da Disciplina 1", "accept": ".pdf", "required": true, "width": 4 } ], [ { "name": "semestre_dis1", "width": 3, "type": "select", "label": "Semestre em que cursou:", "options": [ "1°", "2°" ], "required": true }, { "name": "ano_dis1", "type": "text", "label": "Ano:", "required": true }, { "name": "freq_dis1", "width": 1, "type": "text", "label": "Frequência:", "required": true }, { "name": "nota_dis1", "width": 1, "type": "text", "label": "Nota:", "required": true }, { "name": "credit_dis1", "width": 1, "type": "text", "label": "Créditos", "required": true }, { "name": "cghr_dis1", "type": "text", "label": "Carga horária:", "required": true } ], [ { "name": "coddis2", "type": "text", "label": "Código 2", "required": false, "width": 2 }, { "name": "disciplina2", "type": "text", "label": "Disciplina 2", "required": false }, { "name": "file_dis2", "type": "file", "label": "Ementa da Disciplina 2", "accept": ".pdf", "required": false, "width": 4 } ], [ { "name": "semestre_dis2", "width": 3, "type": "select", "label": "Semestre em que cursou:", "options": [ "1°", "2°" ], "required": false }, { "name": "ano_dis2", "type": "text", "label": "Ano:", "required": false }, { "name": "freq_dis2", "width": 1, "type": "text", "label": "Frequência:", "required": false }, { "name": "nota_dis2", "width": 1, "type": "text", "label": "Nota:", "required": false }, { "name": "credit_dis2", "width": 1, "type": "text", "label": "Créditos", "required": false }, { "name": "cghr_dis2", "type": "text", "label": "Carga horária:", "required": false } ], [ { "name": "coddis3", "type": "text", "label": "Código 3", "required": false, "width": 2 }, { "name": "disciplina3", "type": "text", "label": "Disciplina 3", "required": false }, { "name": "file_dis3", "type": "file", "label": "Ementa da Disciplina 3", "accept": ".pdf", "required": false, "width": 4 } ], [ { "name": "semestre_dis3", "width": 3, "type": "select", "label": "Semestre em que cursou:", "options": [ "1°", "2°" ], "required": false }, { "name": "ano_dis3", "type": "text", "label": "Ano:", "required": false }, { "name": "freq_dis3", "width": 1, "type": "text", "label": "Frequência:", "required": false }, { "name": "nota_dis3", "width": 1, "type": "text", "label": "Nota:", "required": false }, { "name": "credit_dis3", "width": 1, "type": "text", "label": "Créditos", "required": false }, { "name": "cghr_dis3", "type": "text", "label": "Carga horária:", "required": false } ], [ { "name": "coddis4", "type": "disciplina-usp", "label": "Disciplina à requerir equivalência", "required": true, "width": 2 }, { "name": "disciplina4", "type": "text", "label": "Nome da disciplina desejada", "required": true } ], { "name": "place", "type": "hidden", "value": "place_name" }, { "name": "transition", "type": "hidden", "value": "transition_name" } ]', '2025-03-17 19:33:49', '2026-03-26 20:20:12'],
             [2, 'tr_conferencia_inicio', 'equivalencia', null, '[{"name":"retorno","type":"texttextarea","label":"Informações sobre retorno ao aluno","required":true},{"name":"definition_name","type":"hidden","value":"workflowDefinitionName"},{"name":"place","type":"hidden","value":"place_name"},{"name":"transition","type":"hidden","value":"transition_name"}]', '2025-03-17 19:33:49', '2025-03-17 19:33:49'],
             [3, 'tr_conferencia_indica', 'equivalencia', null, '[{"name":"obs","type":"texttextarea","label":"Enviar ao departamento","required":true},{"name":"definition_name","type":"hidden","value":"workflowDefinitionName"},{"name":"place","type":"hidden","value":"place_name"},{"name":"transition","type":"hidden","value":"transition_name"}]', '2025-03-17 19:33:49', '2025-03-17 19:33:49'],
             [4, 'tr_indica_parecerista', 'equivalencia', null, '[{"name":"obs","type":"texttextarea","label":"Nomear o docente a emitir o parecer","required":true},{"name":"definition_name","type":"hidden","value":"workflowDefinitionName"},{"name":"place","type":"hidden","value":"place_name"},{"name":"transition","type":"hidden","value":"transition_name"}]', '2025-03-17 19:33:49', '2025-03-17 19:33:49'],
